pyproject.toml
[tool.poetry]
name = "eversource_scraper"
version = "0.2.0"
description = "Scraper for the Eversource Utilities website"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
mysqlclient = "^2.0.1"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^20.8b1"
pytest = "^6.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
